pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "file-diff-analyzer"
dynamic = ["version"]
description = "A universal library for analyzing differences between files of various formats"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Docus-AI Team", email = "team@docus-ai.com"}
]
maintainers = [
    {name = "Docus-AI Team", email = "team@docus-ai.com"}
]
keywords = ["file-analysis", "text-comparison", "diff-analysis", "document-analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: General",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
requires-python = ">=3.8"
dependencies = [
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
pdf = ["PyPDF2>=3.0.0"]
docx = ["python-docx>=0.8.11"]
excel = ["openpyxl>=3.0.0"]
csv = ["pandas>=1.3.0"]
all = [
    "PyPDF2>=3.0.0",
    "python-docx>=0.8.11",
    "openpyxl>=3.0.0",
    "pandas>=1.3.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/docus-ai/file-diff-analyzer"
Documentation = "https://github.com/docus-ai/file-diff-analyzer/blob/main/README.md"
Repository = "https://github.com/docus-ai/file-diff-analyzer"
"Bug Tracker" = "https://github.com/docus-ai/file-diff-analyzer/issues"

[project.scripts]
file-diff-analyzer = "file_diff_analyzer.cli:main"

[tool.setuptools.dynamic]
version = {attr = "file_diff_analyzer.__version__"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "PyPDF2.*",
    "docx.*",
    "openpyxl.*",
    "pandas.*",
]
ignore_missing_imports = true
